requirements.txt
google-cloud-bigquery>=3.0.0
db-dtypes>=1.0.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
streamlit>=1.28.0
altair>=5.0.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Logging dependencies (Phase 5.1)
structlog>=23.2.0
python-json-logger>=2.0.7

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0
